pyproject.toml
[tool.poetry]
name = "cloner"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pyaudio = "^0.2.14"
ipython = "^8.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
